requirements.txt
Django==3.2.15
djangorestframework==3.14.0
Pillow==9.3.0
requests==2.28.1
spotipy==2.21.0
